requirements.txt
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
selenium==4.16.0
webdriver-manager==4.0.1
python-telegram-bot==20.7
pyTelegramBotAPI==4.15.4 